Re-structured repo in prep for pip release
