+++ EEG-GAN/requirements.txt
@@ -1,0 +1,10 @@
+pandas~=1.3.4
+numpy~=1.21.4
+matplotlib~=3.5.0
+scipy~=1.8.0
+torch~=1.12.1
+torchvision~=0.13.1
+einops~=0.4.1
+torchsummary~=1.5.1
+torchaudio~=0.12.1
+scikit-learn~=1.1.2
