+++ EEG-GAN/trained_models/gansEEGModel.pt
