+++ EEG-GAN/tutorials/images/gan_training_main_help.png
