--- nn_architecture/ttsgan_components.py
@@ -1,175 +1,0 @@
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-from torch import Tensor
-import math
-import numpy as np
-
-from torchvision.transforms import Compose, Resize, ToTensor
-from einops import rearrange, reduce, repeat
-from einops.layers.torch import Rearrange, Reduce
-from torchsummary import summary
-
-
-"""Components for TTS-GAN. Source: https://arxiv.org/abs/2202.02691"""
-
-
-class Gen_TransformerEncoderBlock(nn.Sequential):
-    def __init__(self,
-                 emb_size,
-                 num_heads=5,
-                 drop_p=0.5,
-                 forward_expansion=4,
-                 forward_drop_p=0.5):
-        super().__init__(
-            ResidualAdd(nn.Sequential(
-                nn.LayerNorm(emb_size),
-                MultiHeadAttention(emb_size, num_heads, drop_p),
-                nn.Dropout(drop_p)
-            )),
-            ResidualAdd(nn.Sequential(
-                nn.LayerNorm(emb_size),
-                FeedForwardBlock(
-                    emb_size, expansion=forward_expansion, drop_p=forward_drop_p),
-                nn.Dropout(drop_p)
-            )
-            ))
-
-
-class Gen_TransformerEncoder(nn.Sequential):
-    def __init__(self, depth=8, **kwargs):
-        super().__init__(*[Gen_TransformerEncoderBlock(**kwargs) for _ in range(depth)])
-
-
-class MultiHeadAttention(nn.Module):
-    def __init__(self, emb_size, num_heads, dropout):
-        super().__init__()
-        self.emb_size = emb_size
-        self.num_heads = num_heads
-        self.keys = nn.Linear(emb_size, emb_size)
-        self.queries = nn.Linear(emb_size, emb_size)
-        self.values = nn.Linear(emb_size, emb_size)
-        self.att_drop = nn.Dropout(dropout)
-        self.projection = nn.Linear(emb_size, emb_size)
-
-    def forward(self, x: Tensor, mask: Tensor = None) -> Tensor:
-        queries = rearrange(self.queries(x), "b n (h d) -> b h n d", h=self.num_heads)
-        keys = rearrange(self.keys(x), "b n (h d) -> b h n d", h=self.num_heads)
-        values = rearrange(self.values(x), "b n (h d) -> b h n d", h=self.num_heads)
-        energy = torch.einsum('bhqd, bhkd -> bhqk', queries, keys)  # batch, num_heads, query_len, key_len
-        if mask is not None:
-            fill_value = torch.finfo(torch.float32).min
-            energy.mask_fill(~mask, fill_value)
-
-        scaling = self.emb_size ** (1 / 2)
-        att = F.softmax(energy / scaling, dim=-1)
-        att = self.att_drop(att)
-        out = torch.einsum('bhal, bhlv -> bhav ', att, values)
-        out = rearrange(out, "b h n d -> b n (h d)")
-        out = self.projection(out)
-        return out
-
-
-class ResidualAdd(nn.Module):
-    def __init__(self, fn):
-        super().__init__()
-        self.fn = fn
-
-    def forward(self, x, **kwargs):
-        res = x
-        x = self.fn(x, **kwargs)
-        x += res
-        return x
-
-
-class FeedForwardBlock(nn.Sequential):
-    def __init__(self, emb_size, expansion, drop_p):
-        super().__init__(
-            nn.Linear(emb_size, expansion * emb_size),
-            nn.GELU(),
-            nn.Dropout(drop_p),
-            nn.Linear(expansion * emb_size, emb_size),
-        )
-
-
-class Dis_TransformerEncoderBlock(nn.Sequential):
-    def __init__(self,
-                 emb_size=100,
-                 num_heads=5,
-                 drop_p=0.,
-                 forward_expansion=4,
-                 forward_drop_p=0.,
-                 **kwargs):
-        super().__init__(
-            ResidualAdd(nn.Sequential(
-                nn.LayerNorm(emb_size),
-                MultiHeadAttention(emb_size, num_heads, drop_p),
-                nn.Dropout(drop_p)
-            )),
-            ResidualAdd(nn.Sequential(
-                nn.LayerNorm(emb_size),
-                FeedForwardBlock(
-                    emb_size, expansion=forward_expansion, drop_p=forward_drop_p),
-                nn.Dropout(drop_p)
-            )
-            ))
-
-
-class Dis_TransformerEncoder(nn.Sequential):
-    def __init__(self, depth=8, **kwargs):
-        super().__init__(*[Dis_TransformerEncoderBlock(**kwargs) for _ in range(depth)])
-
-
-class ClassificationHead(nn.Sequential):
-    """Used for the discriminator in the GAN-framework"""
-    def __init__(self, emb_size=100, n_classes=2):
-        super().__init__()
-        self.clshead = nn.Sequential(
-            Reduce('b n e -> b e', reduction='mean'),
-            nn.LayerNorm(emb_size),
-            nn.Linear(emb_size, n_classes)
-        )
-
-    def forward(self, x):
-        out = self.clshead(x)
-        return out
-
-
-class ClassifierHead(nn.Sequential):
-    """Used for binary classification applications"""
-    def __init__(self, emb_size=100, n_classes=2, softmax=False):
-        super().__init__()
-        self.clshead = nn.Sequential(
-            Reduce('b n e -> b e', reduction='mean'),
-            nn.LayerNorm(emb_size),
-            nn.Linear(emb_size, n_classes),
-            nn.Sigmoid(),
-        )
-
-    def forward(self, x):
-        out = self.clshead(x)
-        return out
-
-
-class PatchEmbedding_Linear(nn.Module):
-    # what are the proper parameters set here?
-    def __init__(self, in_channels=21, patch_size=16, emb_size=100, seq_length=1024):
-        # self.patch_size = patch_size
-        super().__init__()
-        # change the conv2d parameters here
-        self.projection = nn.Sequential(
-            Rearrange('b c (h s1) (w s2) -> b (h w) (s1 s2 c)', s1=1, s2=patch_size),
-            nn.Linear(patch_size * in_channels, emb_size)
-        )
-        self.cls_token = nn.Parameter(torch.randn(1, 1, emb_size))
-        self.positions = nn.Parameter(torch.randn((seq_length // patch_size) + 1, emb_size))
-
-    def forward(self, x: Tensor) -> Tensor:
-        b, _, _, _ = x.shape
-        x = self.projection(x)
-        cls_tokens = repeat(self.cls_token, '() n e -> b n e', b=b)
-        # prepend the cls token to the input
-        x = torch.cat([cls_tokens, x], dim=1)
-        # position
-        x += self.positions
-        return x
